requirements.txt
streamlit>=1.29.0
streamlit-option-menu>=0.3.6
google-generativeai>=0.7.2
pdfplumber>=0.10.3
python-docx>=1.1.0
python-dotenv>=1.0.0
jinja2>=3.1.2